final project/requirements.txt
scipy
matplotlib
numpy
pandas
seaborn
plotly
lmfit
imageio
tabulate
scikit-learn
gsheetsdb
youtube_transcript_api
yfinance
patsy
statsmodels
latex
openpyxl
openai
folium
streamlit_folium
scikit-misc
simpy